add fade transitions to scenes, add setup ui
ExperimentController is the main control logic, it is in charge.  It
will persist through all scenes without being reset.  It is a singleton

SetupUIController manages the setup UI, it triggers ExperimentController
to begin the experiment with the settings chosen by the tester

SceneChanger provides methods to switch the scene.  It is a singleton,
but should also be added to every scene; this is so that the reference
to the scene's CameraFader instance can be set once the scene loads;
attempting to create a second instance of SceneChanger merely updates
the CameraFader reference and then deletes the new object from the scene.

CameraFader is a script which is applied to XROrigin's Main Camera in
every scene.  FadeCanvas is a prefab which should be a child of
XROrigin's Main Camera in every scene.  FadeCanvas is a black world
canvas glued to the player's face whose visibility is controlled by
CameraFader.  Every scene needs a CameraFader and a FadeCanvas in order
for transitions to work correctly.

TLDR: every scene needs a SceneChanger, a CameraFader, and a FadeCanvas,
SceneChanger needs a reference to the CameraFader